WiP migrating to Spring Boot 2

+++ src/main/java/com/aidanwhiteley/books/repository/UserRepositoryCustomMethods.java
@@ -4,5 +4,5 @@ import com.aidanwhiteley.books.controller.dtos.ClientRoles;
 
 public interface UserRepositoryCustomMethods {
 	
-	int updateUserRoles(ClientRoles clientRoles);
+	long updateUserRoles(ClientRoles clientRoles);
 }
